requirements_linux.txt
appdirs==1.4.3
asgiref==3.3.1
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.4.3
configparser==5.0.1
cryptography==2.8
defer==1.0.4
Django==3.1.4
django-bootstrap-breadcrumbs==0.9.2
docutils==0.16
idna==2.10
Kivy==2.1.0.dev0
Kivy-Garden==0.1.4
mysqlclient==2.0.3
Pillow==8.1.0
pycairo==1.16.2
pycparser==2.20
Pygments==2.7.4
python-dateutil==2.8.1
python-debian==0.1.36
pytz==2020.4
requests==2.25.1
six==1.14.0
sqlparse==0.4.1
urllib3==1.26.2
